Change magick number for inserting nodes
We spiked a bit on using the ast package Insert function to insert a new
object into a scope, but we found that API to be a lot more cumbersome because
you need to create a new ast.Object for the kind of statement you want to
insert. Clearly this is not a general purpose way for adding code to a file, as
you would with inserting a template.