Remove special files from index
github.com/onsi/ginkgo and github.com/onsi/gomega got into a state where
they were marked as being special files in the index.

This was causing 'git submodule update' to fail because it could not
find a mapping for these special files in the index within .gitmodules.

'git submodule update' is run by 'go get -u', which was failing.

http://stackoverflow.com/questions/4185365/no-submodule-mapping-found-in-gitmodule-for-a-path-thats-not-a-submodule

See the whatchanged for this commit:

160000 ... D Godeps/_workspace/src/github.com/onsi/gomega
160000 ... D Godeps/_workspace/src/github.com/onsi/ginkgo

160000 is the file mode for special files in the index.

Signed-off-by: Kris Hicks <khicks@pivotal.io>